common: eliminate sentinel values and modernize API signatures
Improve type safety across common/ module interfaces:

- hex: change convert_bin_to_hex_new() length param from int to size_t,
  eliminating the possibility of negative lengths.

- file: change file_write()/file_write_append() len param from ssize_t
  to size_t, removing the implicit "pass -1 to mean strlen()" sentinel.
  Callers now compute strlen() explicitly, making buffer lengths visible
  at every call site.  file_get_extension() return type becomes const
  since it returns a pointer into its input.

- fd: change fd_write() len param from ssize_t to size_t, aligning it
  with fd_read() and removing the signed length check.  A buffer length
  is semantically unsigned; this also eliminates implicit narrowing
  conversions at the call sites.

- protobuf: change protobuf_pack_message_new() from an out-parameter
  pattern (uint32_t fn(msg, &ptr)) to returning a protobuf_packed_msg_t
  struct that pairs the buffer pointer with its length.  This keeps the
  two values together and prevents use-after-free of a stale length.

Update all callers in control/, daemon/ accordingly.

+++ common/hex.c
@@ -78,14 +78,14 @@ convert_bin_to_hex(const uint8_t *in, size_t inlen, uint8_t *out, size_t outlen)
 }
 
 char *
-convert_bin_to_hex_new(const uint8_t *bin, int length)
+convert_bin_to_hex_new(const uint8_t *bin, size_t length)
 {
 	size_t len = MUL_WITH_OVERFLOW_CHECK(length, (size_t)2);
 	len = MUL_WITH_OVERFLOW_CHECK(len, sizeof(char));
 	len = ADD_WITH_OVERFLOW_CHECK(len, 1);
 	char *hex = mem_alloc0(len);
 
-	for (int i = 0; i < length; ++i) {
+	for (size_t i = 0; i < length; ++i) {
 		// remember snprintf additionally writes a '0' byte
 		snprintf(hex + i * 2, 3, "%.2x", bin[i]);
 	}
